Fixes #7
Changes the manner in which the database connections are handled. Trying to open an already open connection throws
an error, but closing an already closed connection is fine. Switched the logic around so that we now only open a new
connection if the connection is already closed, but close connections blindly.